[front] enh: Migrate + split swr/conversations to hooks/conversations
