+++ front/components/assistant/conversation/actions/VisualizationActionIframe.tsx
@@ -1,6 +1,6 @@
+import { useVisualizationRetry } from "@app/hooks/conversations";
 import { useSendNotification } from "@app/hooks/useNotification";
 import { clientFetch } from "@app/lib/egress/client";
-import { useVisualizationRetry } from "@app/lib/swr/conversations";
 import datadogLogger from "@app/logger/datadogLogger";
 import type {
   CommandResultMap,
